Adds functionality to the camera
This is a rough prototype for the auto detection of unpackaged foods for the camera. This is not final at all, this approach requires a server to be running on a your network that calls the API.

The reason for this is I have not yet figured out a secure way for the users device to call the api without hardcoding an access key, as well as one without limits. So instead I just decided to require a proxy server that requires gcloud auth for it to work.

Once we find out how to have the user be able to call the API locally and with limits, the functionality will be final.