Adds food recognition function
Adds the ability to recognize non packaged food.

This is a rough prototype for the sake of security. This was the only way I knew how to get this to work without firebase, otherwise we would need to hard code a key or something.

Next week we will work on the firebase implementation, at which point this functionality will be complete.